Changed from Stats to WatchedAttribute so I don't convert a float to a long
Also found out that base game you can't find or its hard to find herbs so I switched to only flowers for potion bases. I deleted vanilla textures and added game: to the domains. Lastly I changed the mortar & pestle to an item because placing it resets the durability and set maxStackSize to 1 because stacks of items used for crafting have all of them degrade rather than one of them